Move `struct Decoder` into own private module
